requirements.txt
asgiref==3.6.0
attrs==23.1.0
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.2
inflection==0.5.1
jsonschema==4.17.3
Markdown==3.4.3
psycopg2-binary==2.9.6
pyrsistent==0.19.3
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
uritemplate==4.1.1
